+++ packages/react/src/hooks/index.ts
@@ -1,7 +1,7 @@
 export * from "./account";
-export * from "./account-mint";
 export * from "./balance";
 export * from "./client";
 export * from "./latest-blockhash";
 export * from "./program-accounts";
 export * from "./signature-statuses";
+export * from "./token-mint";
